Refactor CaReconciler tests
